+++ core/src/main/scala/org/opentorah/store/Terminal.scala
@@ -1,3 +1,0 @@
-package org.opentorah.store
-
-trait Terminal extends Store