+++ core/src/test/scala/org/opentorah/calendar/TextTest.scala
@@ -3,7 +3,7 @@ package org.opentorah.calendar
 import org.opentorah.calendar.jewish.Jewish.{Day, Moment, Month, TimeVector, Year}
 import org.opentorah.calendar.jewish.{Jewish, LeapYearsCycle, Moon, NewYear, Season, Sun}
 import org.opentorah.calendar.roman.Gregorian
-import org.opentorah.metadata.Language
+import org.podval.metadata.Language
 import Week.Day.*
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers