+++ texts/src/main/scala/org/opentorah/texts/rambam/SeferHamitzvosLessons.scala
@@ -1,7 +1,7 @@
 package org.opentorah.texts.rambam
 
-import org.opentorah.metadata.{Name, Named, Names}
-import org.opentorah.store.Selector
+import org.podval.metadata.{Name, Named, Names}
+import org.podval.store.Selector
 import org.podval.xml.{XmlCodec, XmlParser}
 import zio.blocks.schema.{Modifier, Schema}