+++ texts/src/main/scala/org/opentorah/texts/tanach/TanachBook.scala
@@ -1,9 +1,8 @@
 package org.opentorah.texts.tanach
 
-import org.opentorah.metadata.{HasName, Names}
-import org.opentorah.store.{By, Pure}
 import org.opentorah.util.Collections
-import org.opentorah.util.Effects
+import org.podval.metadata.{HasName, Names}
+import org.podval.store.{By, Pure}
 import org.podval.xml.{XmlAst, XmlCodec, XmlError, XmlParser}
 
 trait TanachBook extends HasName, Pure[?] derives CanEqual: // all deriveds are objects; using eq
@@ -47,7 +46,7 @@ private[tanach] object TanachBook:
 
   // unless this is lazy, ZIO deadlocks; see https://github.com/zio/zio/issues/1841
   // ... but it started manifesting only with the switch to ZIO 2.0!
-  private lazy val book2metadata: Map[TanachBook, Metadata] = Collections.mapValues(book2parsed)(metadata => Effects.unsafeRun(metadata.resolve))
+  private lazy val book2metadata: Map[TanachBook, Metadata] = Collections.mapValues(book2parsed)(_.resolve)
 
   def metadata(book: TanachBook): Metadata = book2metadata(book)
 
@@ -60,4 +59,4 @@ private[tanach] object TanachBook:
     val names: Names,
     val chapters: Chapters
   ):
-    def resolve: Effects.IO[Metadata]
+    def resolve: Metadata