+++ texts/src/main/scala/org/opentorah/texts/tanach/Verse.scala
@@ -1,5 +1,5 @@
 package org.opentorah.texts.tanach
 
-import org.opentorah.store.{NumberedStore, Store, Terminal}
+import org.podval.store.{NumberedStore, Store, Terminal}
 
 abstract class Verse(override val number: Int) extends NumberedStore, Terminal